+++ Dockerfile
@@ -23,7 +23,7 @@ RUN yum repolist > /dev/null && \
     cd /opt/app-root/go/src/github.com/sabre1041/istio-pod-network-controller && \
     cp bin/istio-iptables.sh /usr/local/bin/ && \
     dep ensure -vendor-only &&\
-    go build -o bin/istio-pod-network-controller -v main.go && \
+    go build -o bin/istio-pod-network-controller -v cmd/istio-pod-network-controller/main.go && \
     mv bin/istio-pod-network-controller /usr/local/bin && \
     rm -rf ${GOPATH}
